requirements.txt
Flask
prometheus_client
PyYAML
